+++ src/common/sha256.c
