+++ src/gro16/qap.c
