+++ src/tests.c
@@ -1,4 +1,4 @@
-#include "../src/zpie.h"
+#include "zpie.h"
 
 #include "../circuits/mimc.c"
 #include "../circuits/eddsa.c"
@@ -20,18 +20,22 @@ void test_single_constraint()
 
 void test_eddsa_verification()
 {
-    char *B1 = "5299619240641551281634865583518297030282874472190772894086521144482721001553";
-    char *B2 = "16950150798460657717958625567821834550301663161624707787222815936182638968203";
+    point B;
+    B.x = "5299619240641551281634865583518297030282874472190772894086521144482721001553";
+    B.y = "16950150798460657717958625567821834550301663161624707787222815936182638968203";
     
-    char *R1 = "1262948111445225057373438194818763405700457487429548371463214326190311895864"; 
-    char *R2 = "12533500305127747239777484416561675628195562065959201739446841668623540883587";
-    char *A1 = "21629779320182474195265732521833299809982444552305142529409236301104997786342";
-    char *A2 = "9011812445381030664142622066218331845140881847034934166630871421746105699091";
+    eddsaSignature edsig;
+    edsig.R.x = "1262948111445225057373438194818763405700457487429548371463214326190311895864"; 
+    edsig.R.y = "12533500305127747239777484416561675628195562065959201739446841668623540883587";
+    edsig.S = "2674591880888862378688383832785447197125897205360861957116147165712709455207";
+
+    point A;
+    A.x = "21629779320182474195265732521833299809982444552305142529409236301104997786342";
+    A.y = "9011812445381030664142622066218331845140881847034934166630871421746105699091";
+
     char *msg = "1234";
 
-    char *signature = "2674591880888862378688383832785447197125897205360861957116147165712709455207";
-
-    verify_eddsa(B1, B2, R1, R2, A1, A2, msg, signature);
+    verify_eddsa(edsig, B, A, msg);
 }
 
 void test_mimc_hash()
@@ -80,6 +84,21 @@ void test_full_circuits(void)
 
     circuit_selector = 2;
     CU_ASSERT(execute_proof_system() == 1);
+}
+
+//TODO: fix this
+void test_bulletproofs(void)
+{
+    // we init the bulletproofs module, for 2 aggregated proofs of 64 bits
+    bulletproof_init(64, 2);
+
+    // we set some values to prove knowledge of, and compute the proof (../data/bulletproof.params)
+    unsigned char *si[] = {"1234", "5678"};
+    bulletproof_prove(si);
+
+    // we verify the bulletproof (../data/bulletproof.params)
+    if(bulletproof_verify()) printf("Bulletproof verified.\n");
+    else printf("Bulletproof cannot be verified.\n");
 }
 
 int main()
